+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Internal
+
+- remove `MARK` from `Interface` trait
+
 ## [0.5.4] - 2023-01-30
 
 - remove `.parent()` from expand object